+++ tests/Unit/Service/UserServiceTest.php
@@ -30,27 +30,21 @@ use OCA\Workspace\Service\UserService;
 use OCA\Workspace\Service\WorkspaceService;
 use OCP\IGroup;
 use OCP\IGroupManager;
-use OCP\ILogger;
 use OCP\IUser;
 use OCP\IUserSession;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
 
 class UserServiceTest extends TestCase {
-	/** @var IUser */
-	private $user;
-
-	/** @var IGroupManager */
-	private $groupManager;
-
-	/** @var ILogger */
-	private $logger;
-
-	/** @var UserSession */
-	private $userSession;
+	private IGroupManager $groupManager;
+	private IUser $user;
+	private IUserSession $userSession;
+	private LoggerInterface $logger;
+	private WorkspaceService $workspaceService;
 
 	public function setUp(): void {
 		$this->groupManager = $this->createMock(IGroupManager::class);
-		$this->logger = $this->createMock(ILogger::class);
+		$this->logger = $this->createMock(LoggerInterface::class);
 		$this->workspaceService = $this->createMock(WorkspaceService::class);
 
 		// Sets up the user'session
@@ -108,8 +102,8 @@ class UserServiceTest extends TestCase {
 		// Instantiates our service
 		$userService = new UserService(
 			$this->groupManager,
-			$this->logger,
-			$this->userSession);
+			$this->userSession,
+			$this->logger);
 
 		// Runs the method to be tested
 		$result = $userService->isUserGeneralAdmin();
@@ -136,8 +130,8 @@ class UserServiceTest extends TestCase {
 		// Instantiates our service
 		$userService = new UserService(
 			$this->groupManager,
-			$this->logger,
-			$this->userSession);
+			$this->userSession,
+			$this->logger);
 
 		// Runs the method to be tested
 		$result = $userService->isUserGeneralAdmin();
@@ -165,8 +159,8 @@ class UserServiceTest extends TestCase {
 		// Instantiates our service
 		$userService = new UserService(
 			$this->groupManager,
-			$this->logger,
-			$this->userSession);
+			$this->userSession,
+			$this->logger);
 
 		$this->userSession->expects($this->once())
 			->method('getUser')
@@ -204,8 +198,8 @@ class UserServiceTest extends TestCase {
 		// Instantiates our service
 		$userService = new UserService(
 			$this->groupManager,
-			$this->logger,
-			$this->userSession);
+			$this->userSession,
+			$this->logger);
 
 		// Runs the method to be tested
 		$result = $userService->isSpaceManager();
@@ -233,8 +227,8 @@ class UserServiceTest extends TestCase {
 		// Instantiates our service
 		$userService = new UserService(
 			$this->groupManager,
-			$this->logger,
-			$this->userSession);
+			$this->userSession,
+			$this->logger);
 
 		// Runs the method to be tested
 		$result = $userService->isSpaceManagerOfSpace(1);
@@ -262,8 +256,8 @@ class UserServiceTest extends TestCase {
 		// Instantiates our service
 		$userService = new UserService(
 			$this->groupManager,
-			$this->logger,
-			$this->userSession);
+			$this->userSession,
+			$this->logger);
 
 		// Runs the method to be tested
 		$result = $userService->isSpaceManagerOfSpace(1);